Fix composer avatar hydration and harden persisted-store init and login routing
Defer login restoration to onMount so the first client render matches the
logged-out SSR output, fixing the note composer avatar that stayed on the
robohash placeholder. Drive the landing page off the auth stores, gate the
(app) layout's home initialization on the author store, and drop the /post
ssr=false and NoteDialog mount-guard workarounds.

Extract persistedStore into its own dependency-light module (persisted-store.ts)
so calling it during a module's initialization no longer re-enters a
half-initialized WebStorage and crashes SSR with a circular-import "cannot
access before initialization" error.

Wait for login to resolve before the landing and home pages redirect, so a
logged-in user with followees lands on /home instead of the public timeline.

Use kebab-case file names for the new modules (login-state.ts, persisted-store.ts).
