+++ web/src/lib/SeenOnRelayIcon.ts
@@ -1,3 +1,3 @@
-import { persistedStore } from '$lib/WebStorage';
+import { persistedStore } from '$lib/persisted-store';
 
 export const seenOnRelayIcon = persistedStore<boolean>('preference:seen-on-relay-icon', false);
