+++ web/src/lib/ShowVia.ts
@@ -1,3 +1,3 @@
-import { persistedStore } from '$lib/WebStorage';
+import { persistedStore } from '$lib/persisted-store';
 
 export const showVia = persistedStore<boolean>('preference:show-via', false);
