+++ web/src/lib/login-state.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const loginResolved = writable(false);
