+++ web/src/lib/preferences/UserStatus.ts
@@ -1,3 +1,3 @@
-import { persistedStore } from '$lib/WebStorage';
+import { persistedStore } from '$lib/persisted-store';
 
 export const showUserStatus = persistedStore('preference:user-status', true);
