--- web/src/routes/(app)/post/+page.ts
@@ -1,1 +1,0 @@
-export const ssr = false;
