+++ web/src/routes/+page.svelte
@@ -1,25 +1,17 @@
 <script lang="ts">
-	import { run } from 'svelte/legacy';
-
 	import { goto } from '$app/navigation';
-	import type { LayoutData } from './$types';
-	import { followees, pubkey } from '$lib/stores/Author';
+	import { followees, pubkey, author } from '$lib/stores/Author';
+	import { loginResolved } from '$lib/login-state';
 	import Notice from '$lib/components/Notice.svelte';
 	import SplashScreen from './SplashScreen.svelte';
 	import Login from './(app)/Login.svelte';
-
-	interface Props {
-		data: LayoutData;
-	}
-
-	let { data }: Props = $props();
 
 	let homeLink = $derived(
 		$followees.filter((x) => x !== $pubkey).length > 0 ? '/home' : '/public'
 	);
 
-	run(() => {
-		if (data.authenticated) {
+	$effect(() => {
+		if ($author !== undefined) {
 			goto(homeLink);
 		}
 	});
@@ -27,7 +19,7 @@
 
 <Notice />
 
-{#if data.splash}
+{#if !$loginResolved || $pubkey}
 	<SplashScreen />
 {:else}
 	<Login />
